Profile dependent JwtAuthenticationFilters
