+++ src/test/java/org/radarbase/auth/authentication/OAuthHelper.kt
@@ -11,6 +11,7 @@ import org.radarbase.management.domain.Role
 import org.radarbase.management.domain.User
 import org.radarbase.management.repository.UserRepository
 import org.radarbase.management.security.JwtAuthenticationFilter
+import org.radarbase.management.security.JwtAuthenticationFilterImpl
 import org.slf4j.LoggerFactory
 import org.springframework.mock.web.MockHttpServletRequest
 import org.springframework.security.core.Authentication
@@ -127,7 +128,7 @@ object OAuthHelper {
         Mockito.`when`(userRepository.findOneByLogin(ArgumentMatchers.anyString())).thenReturn(
             createAdminUser()
         )
-        return JwtAuthenticationFilter(createTokenValidator(), { auth: Authentication? -> auth })
+        return JwtAuthenticationFilterImpl(createTokenValidator(), { auth: Authentication? -> auth })
     }
 
     /**
